+++ acceptance-tests/dsl/src/main/java/org/hyperledger/besu/tests/acceptance/dsl/condition/bft/AwaitValidatorSetChange.java
@@ -12,10 +12,10 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.tests.acceptance.dsl.condition.ibft2;
+package org.hyperledger.besu.tests.acceptance.dsl.condition.bft;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.hyperledger.besu.tests.acceptance.dsl.transaction.bft.BftTransactions.LATEST;
+import static org.hyperledger.besu.tests.acceptance.dsl.transaction.clique.CliqueTransactions.LATEST;
 
 import org.hyperledger.besu.ethereum.core.Address;
 import org.hyperledger.besu.tests.acceptance.dsl.WaitUtils;
@@ -23,19 +23,23 @@ import org.hyperledger.besu.tests.acceptance.dsl.condition.Condition;
 import org.hyperledger.besu.tests.acceptance.dsl.node.Node;
 import org.hyperledger.besu.tests.acceptance.dsl.transaction.bft.BftTransactions;
 
-public class ExpectValidators implements Condition {
-  private final BftTransactions ibft;
-  private final Address[] validators;
+import java.util.List;
 
-  public ExpectValidators(final BftTransactions ibft, final Address... validators) {
-    this.ibft = ibft;
-    this.validators = validators;
+public class AwaitValidatorSetChange implements Condition {
+
+  private final BftTransactions bft;
+  private final List<Address> initialSigners;
+
+  public AwaitValidatorSetChange(final List<Address> initialSigners, final BftTransactions bft) {
+    this.initialSigners = initialSigners;
+    this.bft = bft;
   }
 
   @Override
   public void verify(final Node node) {
     WaitUtils.waitFor(
+        60,
         () ->
-            assertThat(node.execute(ibft.createGetValidators(LATEST))).containsExactly(validators));
+            assertThat(node.execute(bft.createGetValidators(LATEST))).isNotEqualTo(initialSigners));
   }
 }
