+++ acceptance-tests/dsl/src/main/java/org/hyperledger/besu/tests/acceptance/dsl/condition/bft/BftConditions.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.tests.acceptance.dsl.condition.ibft2;
+package org.hyperledger.besu.tests.acceptance.dsl.condition.bft;
 
 import static org.hyperledger.besu.tests.acceptance.dsl.transaction.clique.CliqueTransactions.LATEST;
 
@@ -31,12 +31,12 @@ import java.util.stream.Collectors;
 
 import com.google.common.collect.ImmutableMap;
 
-public class Ibft2Conditions {
+public class BftConditions {
 
-  private final BftTransactions ibftTwo;
+  private final BftTransactions bft;
 
-  public Ibft2Conditions(final BftTransactions ibftTwo) {
-    this.ibftTwo = ibftTwo;
+  public BftConditions(final BftTransactions bft) {
+    this.bft = bft;
   }
 
   public List<BesuNode> validators(final BesuNode[] nodes) {
@@ -47,7 +47,7 @@ public class Ibft2Conditions {
   }
 
   public ExpectValidators validatorsEqual(final BesuNode... validators) {
-    return new ExpectValidators(ibftTwo, validatorAddresses(validators));
+    return new ExpectValidators(bft, validatorAddresses(validators));
   }
 
   private Address[] validatorAddresses(final BesuNode[] validators) {
@@ -55,15 +55,15 @@ public class Ibft2Conditions {
   }
 
   public Condition awaitValidatorSetChange(final Node node) {
-    return new AwaitValidatorSetChange(node.execute(ibftTwo.createGetValidators(LATEST)), ibftTwo);
+    return new AwaitValidatorSetChange(node.execute(bft.createGetValidators(LATEST)), bft);
   }
 
   public Condition noProposals() {
-    return new ExpectProposals(ibftTwo, ImmutableMap.of());
+    return new ExpectProposals(bft, ImmutableMap.of());
   }
 
   public PendingVotesConfig pendingVotesEqual() {
-    return new PendingVotesConfig(ibftTwo);
+    return new PendingVotesConfig(bft);
   }
 
   public static class PendingVotesConfig {
